requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
attrs==25.4.0
backports.asyncio.runner==1.2.0
exceptiongroup==1.3.0
greenlet==3.2.4
iniconfig==2.3.0
packaging==25.0
playwright==1.55.0
pluggy==1.6.0
pyee==13.0.0
Pygments==2.19.2
pytest==8.4.2
pytest-asyncio==1.2.0
tomli==2.3.0
typing_extensions==4.15.0
playwright-lighthouse>=2.0.0
lighthouse
